requirements.txt
openai
python-dotenv
google-generativeai
langchain-openai
langchain-google-genai
#langchain-core
langchain_ollama
langsmith
langchain_core
streamlit
langchain-aws